% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{confint.ITE}
\alias{confint.ITE}
\title{Confidence Intervals for Bias-corrected ITE Estimators}
\usage{
\method{confint}{ITE}(obj, alpha = 0.05, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{obj}{An object of class `ITE`, a result of a call to `ITE`}

\item{alpha}{Level of significance to construct confidence interval (default=0.05)}

\item{alternative}{Indicates the alternative hypothesis to construct confidence interval and must be one of "two.sided" (default), "less", or "greater".}
}
\value{
A matrix with columns giving lower and upper confidence limits for bias-corrected
estimators.
}
\description{
Computes confidence intervals for bias-corrected estimators; Each
row corresponds to a loading.
}
\examples{
\dontrun{
##-- Continuing the ITE(.) example:
ci = confint(Est)
ci
}
}
