% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{confint.LF}
\alias{confint.LF}
\title{Confidence Intervals for Bias-corrected LF Estimators}
\usage{
\method{confint}{LF}(obj, alpha = 0.05, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{obj}{An object of class `LF`, a result of a call to `LF`}

\item{alpha}{Level of significance to construct confidence interval (default=0.05)}

\item{alternative}{Indicates the alternative hypothesis to construct confidence interval and must be one of "two.sided" (default), "less", or "greater".}
}
\value{
A matrix with columns giving lower and upper confidence limits for bias-corrected
estimators.
}
\description{
Computes confidence intervals for bias-corrected estimators; Each
row corresponds to a loading.
}
\examples{
\dontrun{
##-- Continuing the LF(.) example:
ci = confint(Est)
ci
}
}
