% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_linear.R
\name{Direction_fixedtuning}
\alias{Direction_fixedtuning}
\title{Constructs the projection direction with fixed tuning parameter for estimating functional in high-dimensional linear regression model}
\usage{
Direction_fixedtuning(Xc, loading, mu = NULL)
}
\arguments{
\item{Xc}{Design matrix, of dimension nvar(\eqn{n})xnobs(\eqn{p})}

\item{loading}{observation vector in the linear functional, of length \eqn{p}}

\item{mu}{Tuning parameter in construction of projection direction}
}
\value{
\item{proj}{The projection direction}
}
\description{
Constructs the projection direction with fixed tuning parameter for estimating functional in high-dimensional linear regression model
}
\examples{
X = matrix(sample(-2:2,100*400,replace = TRUE),nrow=100,ncol=400)
Direction_fixedtuning(X,loading=c(1,rep(0,399)),mu=2)

}
