% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_logistic.R
\name{Direction_fixedtuning_logistic}
\alias{Direction_fixedtuning_logistic}
\title{Constructs the projection direction with fixed tuning parameter in high dimensional logistic regression}
\usage{
Direction_fixedtuning_logistic(X, loading, mu = NULL, weight, f_prime)
}
\arguments{
\item{X}{Design matrix, of dimension \eqn{n} x \eqn{p}}

\item{loading}{Loading, of length \eqn{p}}

\item{mu}{The tuning parameter used in the construction of the projection direction}

\item{weight}{The weight vector, of length \eqn{n}, used in correcting the plug-in estimator}

\item{f_prime}{The first derivative of the logit function at the plug-in LASSO estimate of the linear functional}
}
\value{
\item{proj}{The projection direction}
}
\description{
Constructs the projection direction with fixed tuning parameter in high dimensional logistic regression
}
\examples{
n = 50
p = 400
X = matrix(sample(-2:2,n*p,replace = TRUE),nrow=n,ncol=p)
y = rbinom(n,1,0.5)
col.norm <- 1/sqrt((1/n)*diag(t(X)\%*\%X)+0.0001);
Xnor <- X \%*\% diag(col.norm);
fit = glmnet::cv.glmnet(Xnor, y, alpha=1,family = "binomial")
htheta <- as.vector(coef(fit, s = "lambda.min"))
support<-(abs(htheta)>0.001)
Xb <- cbind(rep(1,n),Xnor);
Xc <- cbind(rep(1,n),X);
col.norm <- c(1,col.norm);
pp <- (p+1);
xnew = c(1,rep(0,(p-1)))
loading=rep(0,pp)
loading[1]=1
loading[-1]=xnew
htheta <- htheta*col.norm;
htheta <- as.vector(htheta)
f_prime <- exp(Xc\%*\%htheta)/(1+exp(Xc\%*\%htheta))^2
Direction_fixedtuning_logistic(Xc,loading,mu=2,weight=rep(1,50),f_prime)
}
