% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_linear.R
\name{Direction_searchtuning}
\alias{Direction_searchtuning}
\title{Searches for the best step size and computes the projection direction in high dimensional linear regression}
\usage{
Direction_searchtuning(X, loading, mu = NULL, resol = 1.5, maxiter = 10)
}
\arguments{
\item{X}{Design matrix, of dimension \eqn{n} x \eqn{p}}

\item{loading}{Loading, of length \eqn{p}}

\item{mu}{Tuning parameter in construction of projection direction}

\item{resol}{Resolution or the factor by which \code{mu} is increased/decreased to obtain the smallest \code{mu}
that gives convergence of the optimization problem for constructing the projection direction (default = 1.5)}

\item{maxiter}{Maximum number of steps along which \code{mu} is increased/decreased to obtain the smallest \code{mu}
that gives convergence of the optimization problem for constructing the projection direction (default = 10)}
}
\value{
\item{proj}{The projection direction}
}
\description{
Searches for the best step size and computes the projection direction in high dimensional linear regression
}
\examples{
n = 100
p = 400
X = matrix(sample(-2:2,n*p,replace = TRUE),nrow = n,ncol = p)
Direction_searchtuning(X,loading=c(1,rep(0,(p-1))))
}
