% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_LF.R
\name{GLM_LF}
\alias{GLM_LF}
\title{Inference for linear combination of the regression vector in high dimensional generalized linear regression}
\usage{
GLM_LF(
  X,
  y,
  model = "linear",
  loading,
  intercept.loading = TRUE,
  intercept = TRUE,
  init.coef = NULL,
  lambda = NULL,
  mu = NULL,
  step = NULL,
  resol = 1.5,
  maxiter = 6,
  b0 = 0,
  alpha = 0.05,
  verbose = TRUE
)
}
\arguments{
\item{X}{Design matrix, of dimension \eqn{n} x \eqn{p}}

\item{y}{Outcome vector, of length \eqn{n}}

\item{model}{The high dimensional regression model, either \code{linear} or \code{logistic} or \code{logistic alternative} or \code{probit} or \code{inverse t1}}

\item{loading}{Loading, of length \eqn{p}}

\item{intercept.loading}{Should intercept be included for the \code{loading} (default = \code{TRUE})}

\item{intercept}{Should intercept be fitted for the initial estimator (default = \code{TRUE})}

\item{init.coef}{Initial estimator of the regression vector (default = \code{NULL})}

\item{lambda}{The tuning parameter in the construction of \code{init.coef} (default = \code{NULL})}

\item{mu}{The dual tuning parameter used in the construction of the projection direction (default = \code{NULL})}

\item{step}{The step size used to compute \code{mu}; if set to \code{NULL} it is
computed to be the number of steps (< \code{maxiter}) to obtain the smallest \code{mu}
such that the dual optimization problem for constructing the projection direction converges (default = \code{NULL})}

\item{resol}{The factor by which \code{mu} is increased/decreased to obtain the smallest \code{mu}
such that the dual optimization problem for constructing the projection direction converges (default = 1.5)}

\item{maxiter}{Maximum number of steps along which \code{mu} is increased/decreased to obtain the smallest \code{mu}
such that the dual optimization problem for constructing the projection direction converges (default = 6)}

\item{b0}{The null value to be tested against                                                              #removed}

\item{alpha}{Level of significance to test the null hypothesis which claims that the linear combination of the regression coefficients
is less than or equal to \code{b0} (default = 0.05)                                                        #removed}

\item{verbose}{Should inetrmediate message(s) be printed (default = \code{TRUE})}
}
\value{
\item{prop.est}{The bias-corrected estimator for the linear combination of regression coefficients}
\item{se}{The standard error of the bias-corrected estimator}
\item{CI}{The confidence interval for the linear combination}                                                     #removed
\item{decision}{\code{decision}\eqn{=1} implies the linear combination is above \code{b0} \eqn{\newline}          #removed
\code{decision}\eqn{=0} implies the linear combination is not above \code{b0}}
\item{proj}{The projection direction, of length \eqn{p}}
\item{plug.in}{The plug-in estimator for the linear combination}
}
\description{
Computes the bias-corrected estimator of thelinearcombination of the regression vector for the high dimensional generalized linear regression and the corresponding standard error.
}
\examples{

\donttest{
sp = 20
n = 400
p = 800
f = function(x){
pnorm(x)
}
sig1 = toeplitz(seq(0.6, 0,length.out = p/10))
Sig = Matrix::bdiag(rep(list(sig1),10))+diag(rep(0.4,p))
set.seed(1203)
X = MASS::mvrnorm(n, mu=rep(0,p), Sigma=Sig)
b = rep(0,p)
b[1:sp] = rep(c(0.4,-0.4), sp/2)
prob = f(X \%*\% b)
y = array(dim = 1)
for(i in 1:n){
set.seed(i)
y[i] = rbinom(1,1,prob[i])
}
Est = SIHR::GLM_LF(X = X, y = y, loading = c(1,rep(0,799)), model = "probit", intercept = FALSE)
}
}
