% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group_Test_Covariance_linear.R
\name{Group_Test_Cov}
\alias{Group_Test_Cov}
\title{Inference for quadratic functional in high-dimensional linear regression}
\usage{
Group_Test_Cov(
  X,
  y,
  test.set,
  tau.vec = NULL,
  init.Lasso = NULL,
  lambda = NULL,
  intercept = FALSE,
  mu = NULL,
  step = NULL,
  resol = 1.5,
  maxiter = 6
)
}
\arguments{
\item{X}{Design matrix, of dimensions nobs(n) x nvar(p)}

\item{y}{Response variable}

\item{test.set}{set of indices, G in \eqn{\beta_G^{\top}\Sigma_{G,G}\beta_G}}

\item{tau.vec}{Vector of enlargement factors for asymptotic variance of the estimator \eqn{\widehat{\beta_G^{\top}\Sigma_{G,G}\beta_G}} to handle super-efficiency}

\item{init.Lasso}{initial LASSO estimator \eqn{\widehat{\beta}} (default = \code{NULL})}

\item{lambda}{Tuning parameter \eqn{\lambda} used in construction of initial LASSO estimator \eqn{\widehat{\beta}} if \code{init.Lasso = NULL}}

\item{intercept}{Should intercept(s) be fitted (default = \code{FALSE})}

\item{mu}{Tuning parameter \eqn{\mu} in construction of projection direction \eqn{\widehat{u}}}

\item{step}{Number of steps (< \code{maxiter}) to obtain the smallest \code{mu} that gives convergence of the
optimization problem for constructing the projection direction \eqn{\widehat{u}} (default = \code{NULL})}

\item{resol}{Resolution or the factor by which \code{mu} is increased/decreased to obtain the smallest \code{mu}
that gives convergence of the optimization problem for constructing the projection direction \eqn{\widehat{u}}}

\item{maxiter}{Maximum number of steps along which \code{mu} is increased/decreased to obtain the smallest \code{mu}
that gives convergence of the optimization problem for constructing the projection direction \eqn{\widehat{u}}}
}
\value{
\item{prop.est}{The bias-corrected estimator \eqn{\widehat{\beta_G^{\top}\Sigma_{G,G}\beta_G}}}
\item{sigma}{Estimate of the variance of error term \eqn{\epsilon} in the linear regression model}
\item{se}{Standard error of \eqn{\widehat{\beta_G^{\top}\Sigma_{G,G}\beta_G}}}
\item{plug.in}{Plug-in LASSO estimator \eqn{\widehat{\beta}_{G}^{\top} \widehat{\Sigma}_{G, G} \widehat{\beta}_{G}}}
}
\description{
Computes the bias-corrected estimator of \eqn{\beta_G^{\top}\Sigma_{G,G}\beta_G} for the following high-dimensional
linear regression model \deqn{y=X\beta+\epsilon}
as \deqn{\widehat{\beta_G^{\top}\Sigma_{G,G}\beta_G}=\widehat{\beta}_{G}^{\top} \widehat{\Sigma}_{G, G} \widehat{\beta}_{G}+\frac{2}{n} \widehat{u}^{\top} X^{\top}(y-X \widehat{\beta})}
where \eqn{\Sigma} is the population design covariance matrix and \eqn{G\subset{1,2,...,p}}
}
\examples{
X = matrix(sample(-2:2,100*400,replace = TRUE),nrow=100,ncol=400)
beta = (1:400)/25
y = X\%*\%beta + rnorm(100,0,1)
Group_Test_Cov(X = X, y = y, test.set=c(30:50))
}
