% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inference.R
\name{inf}
\alias{inf}
\title{Confidence interval and hypothesis testing in high-dimensional generalized linear regression}
\usage{
inf(object, alpha = 0.05, b0 = 0)
}
\arguments{
\item{object}{an object of class \code{LF}, \code{ITE} or \code{QF} for which inference is desired}

\item{alpha}{Level of significance to test the null hypothesis}

\item{b0}{The null value to be tested against}
}
\value{
\item{CI}{The confidence interval for the target quantity}
\item{decision}{\code{decision}\eqn{=1} implies the target quantity is above \code{b0} \eqn{\newline}
\code{decision}\eqn{=0} implies the target quantity is above \code{b0}}
}
\description{
Constructs the confidence interval for the target quantity and test whether it is above \code{b0} or not
}
\examples{
n = 100
p = 400
A1gen <- function(rho,p){
  A1=matrix(0,p,p)
  for(i in 1:p){
    for(j in 1:p){
      A1[i,j]<-rho^(abs(i-j))
    }
  }
  A1
}
mu <- rep(0,p)
mu[1:5] <- c(1:5)/5
rho = 0.5
Cov <- (A1gen(rho,p))/2
beta <- rep(0,p)
beta[1:10] <- c(1:10)/5
X <- MASS::mvrnorm(n,mu,Cov)
y = X\%*\%beta + rnorm(n)
loading <- MASS::mvrnorm(1,rep(0,p),Cov)
Est = SIHR::GLM_LF(X = X, y = y, loading = loading)
CI = inf(Est)$CI
dec = inf(Est)$decision
}
