% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_Inference_linear.R
\name{LF_Inference}
\alias{LF_Inference}
\title{Inference for linear functional in the high-dimensional linear regression model}
\usage{
LF_Inference(
  X,
  y,
  loading,
  lambda = NULL,
  intercept = FALSE,
  mu = NULL,
  step = NULL,
  resol = 1.5,
  maxiter = 10
)
}
\arguments{
\item{X}{Design matrix, of dimensions nobs(n) x nvar(p)}

\item{y}{Response variable}

\item{loading}{\eqn{x_{*}} in \eqn{x_{*}^{\top}\beta}}

\item{lambda}{Tuning parameter \eqn{\lambda} in construction of LASSO estimator \eqn{\widehat{\beta}}}

\item{intercept}{Should intercept(s) be fitted (default = \code{FALSE})}

\item{mu}{Tuning parameter \eqn{\mu} in construction of projection direction \eqn{\widehat{u}}}

\item{step}{Number of steps (< \code{maxiter}) to obtain the smallest \code{mu} that gives convergence of the
optimization problem for constructing the projection direction \eqn{\widehat{u}} (default = \code{NULL})}

\item{resol}{Resolution or the factor by which \code{mu} is increased/decreased to obtain the smallest \code{mu}
that gives convergence of the optimization problem for constructing the projection direction \eqn{\widehat{u}}}

\item{maxiter}{Maximum number of steps along which \code{mu} is increased/decreased to obtain the smallest \code{mu}
that gives convergence of the optimization problem for constructing the projection direction \eqn{\widehat{u}}}
}
\value{
\item{prop.est}{The bias-corrected estimator \deqn{\widehat{x_{*}^{\top}\beta}}}
\item{sigma}{Estimate of the variance of error term \eqn{\epsilon} in the linear regression model}
\item{se}{Standard error of \eqn{\widehat{x_{*}^{\top}\beta}}}
\item{proj}{Projection direction \eqn{\widehat{u}} in \eqn{\widehat{x_*^{\top}\beta}}}
\item{step}{Number of steps (< \code{maxiter}) to obtain the smallest \code{mu} that gives convergence of the
optimization problem for constructing the projection direction \eqn{\widehat{u}}}
\item{plug.in}{Plug-in LASSO estimator \eqn{x_{*}^{\top}\widehat{\beta}}}
}
\description{
Computes the bias corrected estimator for \eqn{x_*^{\top}\beta} for the following high-dimensional linear regression model
\deqn{y=X\beta+\epsilon} as \deqn{\widehat{x_{*}^{\top}\beta}=x_{*}^{\top}\widehat{\beta}+\widehat{u}^{\top}\frac{1}{n}\sum_{i=1}^{n}X_{i.}\left(y_i-X_{i.}^{\top}\beta\right)}
and the corresponding standard error.
}
\examples{
X = matrix(sample(-2:2,100*400,replace = TRUE),nrow=100,ncol=400)
beta = (1:400)/25
y = X\%*\%beta + rnorm(100,0,1)
LF_Inference(X = X, y = y, loading = c(1,rep(0,399)), intercept = TRUE)
}
