% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_Inference_logistic.R
\name{LF_Inference_logistic}
\alias{LF_Inference_logistic}
\title{Inference for linear functional in high dimensional logistic regression}
\usage{
LF_Inference_logistic(
  X,
  y,
  loading,
  lambda = NULL,
  intercept = FALSE,
  mu = NULL,
  step = NULL,
  resol = 1.5,
  maxiter = 6
)
}
\arguments{
\item{X}{input matrix, of dimension nobs x nvar; each row is an observation vector}

\item{y}{response variable}

\item{loading}{the vector of covariates in the linear functional to be estimated}

\item{lambda}{tuning parameter}

\item{intercept}{Should intercept(s) be fitted (default = \code{FALSE})}

\item{mu}{tuning parameter for construction of projection direction}

\item{step}{number of steps (< \code{maxiter}) to obtain the smallest \code{mu} that gives convergence}

\item{resol}{resolution or the factor by which \code{mu} is increased/decreased till smallest with convergence}

\item{maxiter}{maximum number of steps along which we increase/decrease \code{mu} till smallest with convergence}
}
\value{
bias corrected linear  functional estimate and corresponding standard error for high dimensional logistic regression model
}
\description{
Inference for linear functional in high dimensional logistic regression
}
\examples{
LF_Inference_logistic(X = matrix(sample(-2:2,100*400,replace = TRUE),nrow=100,ncol=400),
                      y = rbinom(100,1,0.5), loading = c(1,rep(0,399)), intercept = TRUE)
}
