% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weight_fn_LF_logistic.R
\name{LF_Inference_weight}
\alias{LF_Inference_weight}
\title{Inference for linear functional in the high-dimensional logistic regression model}
\usage{
LF_Inference_weight(
  X,
  y,
  loading,
  lambda = NULL,
  intercept = FALSE,
  mu = NULL,
  step = NULL,
  weight,
  resol = 1.5,
  maxiter = 6
)
}
\arguments{
\item{X}{Design matrix, of dimensions nobs(n) x nvar(p)}

\item{y}{Response vector}

\item{loading}{\eqn{x_{*}} in \eqn{x_{*}^{\top}\beta}}

\item{lambda}{Tuning parameter \eqn{\lambda} in construction of LASSO estimator \eqn{\widehat{\beta}}}

\item{intercept}{Should intercept(s) be fitted (default = \code{FALSE})}

\item{mu}{Tuning parameter \eqn{\mu} in construction of projection direction \eqn{\widehat{u}}}

\item{step}{Number of steps (< \code{maxiter}) to obtain the smallest \code{mu} that gives convergence of the
optimization problem for constructing the projection direction \eqn{\widehat{u}} (default = \code{NULL})}

\item{weight}{The vector of weights \eqn{(W_i)_{i=1}^{n}} used in correcting the plug-in estimator \eqn{x_{*}^{\top}\widehat{\beta}}}

\item{resol}{Resolution or the factor by which \code{mu} is increased/decreased to obtain the smallest \code{mu}
that gives convergence of the optimization problem for constructing the projection direction \eqn{\widehat{u}}}

\item{maxiter}{Maximum number of steps along which \code{mu} is increased/decreased to obtain the smallest \code{mu}
that gives convergence of the optimization problem for constructing the projection direction \eqn{\widehat{u}}}
}
\value{
\item{prop.est}{The bias corrected estimator \deqn{\widehat{x_{*}^{\top}\beta}}.}
\item{se}{Standard error of \eqn{\widehat{x_{*}^{\top}\beta}}}
\item{proj}{Projection direction \eqn{\widehat{u}} in \eqn{\widehat{x_*^{\top}\beta}}}
\item{step}{Number of steps (< \code{maxiter}) to obtain the smallest \code{mu} that gives convergence of the
optimization problem for constructing the projection direction \eqn{\widehat{u}}}
\item{plug.in}{Plug-in LASSO estimator \eqn{x_{*}^{\top}\widehat{\beta}}}
\item{support}{Estimated support for \eqn{\beta} (using LASSO estimate \eqn{\widehat{\beta}})}
}
\description{
Computes the bias corrected estimator for \eqn{x_*^{\top}\beta} for the following high-dimensional linear regression model
\deqn{y=\frac{e^{X\beta}}{1+e^{X\beta}}+\epsilon} as \deqn{\widehat{x_{*}^{\top}\beta}=x_{*}^{\top}\widehat{\beta}+\widehat{u}^{\top}\frac{1}{n}\sum_{i=1}^{n}W_{i}\left(y_i-\frac{e^{X_{i.}^{\top}\widehat{\beta}}}{1+e^{X_{i.}^{\top}\widehat{\beta}}}\right)}
and the corresponding standard error.
}
\examples{
LF_Inference_weight(X = matrix(sample(-2:2,100*400,replace = TRUE),nrow=100,ncol=400),
                      y = rbinom(100,1,0.5), loading = c(1,rep(0,399)),
                      intercept = TRUE, weight = rep(1,100))
}
