% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{print.summary.CATE}
\alias{print.summary.CATE}
\title{Printing Summarizing CATE}
\usage{
\method{print}{summary.CATE}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{An object of class `summary.CATE`, a result of a call to `summary.CATE`}

\item{digits}{The number of digits to use when printing}

\item{...}{arguments to pass down}
}
\description{
`print` method for class `summary.CATE`
}
\examples{
\dontrun{
#' ##-- Continuing the CATE(.) example:
summary(Est)
}
}
\keyword{internal}
