% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{print.summary.Dist}
\alias{print.summary.Dist}
\title{Printing summarizing Dist}
\usage{
\method{print}{summary.Dist}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{An object of class `summary.Dist`, a result of a call to `summary.Dist`}

\item{digits}{The number of digits to use when printing}

\item{...}{arguments to pass down}
}
\description{
`print` method for class `summary.Dist`
}
\examples{
\dontrun{
#' ##-- Continuing the Dist(.) example:
summary(Est)
}
}
\keyword{internal}
