% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{print.summary.InnProd}
\alias{print.summary.InnProd}
\title{Printing summarizing InnProd}
\usage{
\method{print}{summary.InnProd}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{An object of class `summary.InnProd`, a result of a call to `summary.InnProd`}

\item{digits}{The number of digits to use when printing}

\item{...}{arguments to pass down}
}
\description{
`print` method for class `summary.InnProd`
}
\examples{
\dontrun{
#' ##-- Continuing the InnProd(.) example:
summary(Est)
}
}
\keyword{internal}
