% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{print.summary.QF}
\alias{print.summary.QF}
\title{Summarizing QF}
\usage{
\method{print}{summary.QF}(obj, digits = 5)
}
\arguments{
\item{obj}{An object of class `summary.QF`, a result of a call to `summary.QF`}

\item{digits}{The number of digits to use when printing}
}
\description{
`summary` method for class `QF`
}
\examples{
\dontrun{
#' ##-- Continuing the QF(.) example:
summary(Est)
}
}
