% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QF.R
\name{QF}
\alias{QF}
\title{Inference for quadratic functional in high dimensional linear regression model}
\usage{
QF(
  X,
  y,
  test.set,
  A = "Sigma",
  init.Lasso = NULL,
  tau.vec = NULL,
  lambda = NULL,
  intercept = TRUE,
  mu = NULL,
  step = NULL,
  resol = 1.5,
  maxiter = 10
)
}
\arguments{
\item{X}{Design matrix, of dimension \eqn{n} x \eqn{p}}

\item{y}{Outcome vector, of length \eqn{n}}

\item{test.set}{The set of indices, \code{G} in the quadratic functional}

\item{A}{The matrix A in the quadratic functional, of dimension \eqn{p}x\eqn{p} (either the population covariance matrix \code{Sigma} or a known matrix of suitable dimension ; default = \code{Sigma})}

\item{init.Lasso}{Initial LASSO estimator for the regression vector (default = \code{NULL})}

\item{tau.vec}{The vector of enlargement factors for asymptotic variance of the bias-corrected estimator to handle super-efficiency (default = \code{NULL})}

\item{lambda}{The tuning parameter used in construction of initial LASSO estimator of the regression vector if \code{init.Lasso = NULL} (default = \code{NULL})}

\item{intercept}{Should intercept(s) be fitted (default = \code{TRUE})}

\item{mu}{The tuning parameter in construction of the projection direction (default = \code{NULL})}

\item{step}{Number of steps (< \code{maxiter}) to obtain the smallest \code{mu} that gives convergence of the
optimization problem for constructing the projection direction (default = \code{NULL})}

\item{resol}{Resolution or the factor by which \code{mu} is increased/decreased to obtain the smallest \code{mu}
that gives convergence of the optimization problem for constructing the projection direction (default = 1.5)}

\item{maxiter}{Maximum number of steps along which \code{mu} is increased/decreased to obtain the smallest \code{mu}
that gives convergence of the optimization problem for constructing the projection direction (default = 10)}
}
\value{
\item{prop.est}{The bias-corrected estimator of the quadratic functional restricted to \code{G}}
\item{sigma}{Estimate of the error variance in the linear regression model}
\item{se}{Standard error of the bias-corrected estimator}
\item{plug.in}{Plug-in LASSO estimator for the quadratic functional restricted to \code{G}}
}
\description{
Computes the bias-corrected estimator of the quadratic functional \eqn{\beta_G^{\top}A\beta_G} for the high dimensional linear regression model \eqn{Y_i = X_i^{\top}\beta + \epsilon} and the corresponding standard error.
}
\examples{
n = 100
p = 400
X = matrix(sample(-2:2,n*p,replace = TRUE),nrow=n,ncol=p)
beta = (1:p)/25
y = X\%*\%beta + rnorm(n,0,1)
test.set =c(30:50)
QF(X = X, y = y, test.set)

}
\references{
\insertRef{grouplin}{FIHR}
}
