% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSLasso.R
\name{SSLasso}
\alias{SSLasso}
\title{Compute confidence intervals and p-values.}
\usage{
SSLasso(
  X,
  y,
  alpha = 0.05,
  lambda = NULL,
  mu = NULL,
  intercept = TRUE,
  resol = 1.3,
  maxiter = 50,
  threshold = 0.01,
  verbose = TRUE
)
}
\arguments{
\item{X}{Design matrix, of dimensions nobs(n) x nvar(p)}

\item{y}{Response vector}

\item{alpha}{Significance level}

\item{lambda}{Lasso regularization parameter (if \code{NULL}, fixed by square-root LASSO)}

\item{mu}{\eqn{\ell_{\infty}} constraint on M (if \code{NULL}, searches)}

\item{intercept}{Should intercept(s) be fitted (default = \code{FALSE})}

\item{resol}{Step parameter for the function that computes M}

\item{maxiter}{Iteration parameter for computing M}

\item{threshold}{Tolerance criterion for computing M}

\item{verbose}{verbose}
}
\value{
\item{noise.sd}{Estimate of the noise standard deviation}
\item{norm0}{Estimate of the number of 'significant' coefficients}
\item{coef}{Lasso estimated coefficients}
\item{unb.coef}{Unbiased coefficient estimates}
\item{low.lim}{Lower limits of confidence intervals}
\item{up.lim}{upper limit of confidence intervals}
\item{pvals}{p-values for the coefficients}
}
\description{
Compute confidence intervals and p-values.
}
\examples{
X = matrix(sample(-2:2,100*400,replace = TRUE),nrow=100,ncol=400)
beta = (1:400)/25
y = X\%*\%beta + rnorm(100,0,1)
SSLasso(X = X, y = y)
}
