% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary.CATE}
\alias{summary.CATE}
\title{Summarizing CATE}
\usage{
\method{summary}{CATE}(object, ...)
}
\arguments{
\item{object}{An object of class `CATE`, a result of a call to `CATE`}

\item{...}{arguments to pass down}
}
\value{
The function `summary.CATE` computes and returns a list of summary statistics.
\item{output.est}{a \eqn{ncol(loading.mat)} x 7 matrix with columns for the loading,
plugin(biased) estimators, bias-corrected estimators, its standard error, z-statistic,
corresponding (two-sided) p-value and significance stars; Each row corresponds to each loading.}
}
\description{
`summary` method for class `CATE`
}
\examples{
\dontrun{
##-- Continuing the CATE(.) example:
sEst = summary(Est)
sEst
}
}
\keyword{internal}
