% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary.Dist}
\alias{summary.Dist}
\title{Summarizing Dist}
\usage{
\method{summary}{Dist}(object, ...)
}
\arguments{
\item{object}{An object of class `Dist`, a result of a call to `Dist`}

\item{...}{arguments to pass down}
}
\value{
The function `summary.Dist` computes and returns a list of summary statistics.
\item{output.est}{A 6-dimensional vector with elements for plugin(biased) estimators, bias-corrected estimators, its standard error, z-statistic,
corresponding (two-sided) p-value and significance stars}
}
\description{
`summary` method for class `Dist`
}
\examples{
\dontrun{
##-- Continuing the Dist(.) example:
sEst = summary(Est)
sEst
}
}
\keyword{internal}
