% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary.InnProd}
\alias{summary.InnProd}
\title{Summarizing InnProd}
\usage{
\method{summary}{InnProd}(object, ...)
}
\arguments{
\item{object}{An object of class `InnProd`, a result of a call to `InnProd`}

\item{...}{arguments to pass down}
}
\value{
The function `summary.InnProd` computes and returns a list of summary statistics.
\item{output.est}{A 6-dimensional vector with elements for plugin(biased) estimators, bias-corrected estimators, its standard error, z-statistic,
corresponding (two-sided) p-value and significance stars}
}
\description{
`summary` method for class `InnProd`
}
\examples{
\dontrun{
##-- Continuing the InnProd(.) example:
sEst = summary(Est)
sEst
}
}
\keyword{internal}
