% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary.QF}
\alias{summary.QF}
\title{Summarizing QF}
\usage{
\method{summary}{QF}(obj, alpha = 0.05, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{obj}{An object of class `QF`, a result of a call to `QF`}

\item{alpha}{Level of significance to construct confidence interval}

\item{alternative}{indicates the alternative hypothesis to construct confidence interval and must be one of "two.sided" (default), "less", or "greater".}
}
\value{
The function `summary.QF` computes and returns a list of summary statistics
of LF given `obj`
\item{output.est}{a \eqn{length(tau.vec)} x 5 matrix with columns for the
plugin(biased) estimators, bias-corrected estimators, its standard error, z-statistic and
corresponding (two-sided) p-value; Each row corresponds to each tau.}
\item{output.ci}{a \eqn{length(tau.vc)} x 2 matrix with columns for lower and upper
of (two-sided) confidence intervals; Each row corresponds to each tau.}
}
\description{
`summary` method for class `QF`
}
\examples{
\dontrun{
##-- Continuing the QF(.) example:
sEst = summary(Est)
sEst
}
}
