% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{ci}
\alias{ci}
\title{Confidence Intervals for Bias-corrected Estimators}
\usage{
ci(
  object,
  probability = FALSE,
  alpha = 0.05,
  alternative = c("two.sided", "less", "greater"),
  ...
)
}
\arguments{
\item{object}{An object of class}

\item{probability}{Whether returns CI with probability transformation or not (default=\code{FALSE})}

\item{alpha}{Level of significance to construct confidence interval (default=0.05)}

\item{alternative}{Indicates the alternative hypothesis to construct confidence interval and must be one of "two.sided" (default), "less", or "greater".}

\item{...}{arguments to pass down}
}
\description{
generic function
}
\keyword{internal}
