% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projection_Directions.R
\name{Direction_fixedtuning}
\alias{Direction_fixedtuning}
\title{Constructs the projection direction with fixed tuning parameter in high dimensional regression}
\usage{
Direction_fixedtuning(
  X,
  loading,
  mu = NULL,
  model = "linear",
  weight = NULL,
  deriv.vec = NULL
)
}
\arguments{
\item{X}{Design matrix, of dimension \eqn{n} x \eqn{p}}

\item{loading}{Loading, of length \eqn{p}}

\item{mu}{The tuning parameter used in the construction of the projection direction}

\item{model}{The high dimensional regression model, either \code{linear} or \code{logistic}}

\item{weight}{The weight vector of length \eqn{n}, used in correcting the plug-in estimator ; to be supplied if \code{model="logistic"} (default=NULL)}

\item{deriv.vec}{The first derivative vector of the logit function at \eqn{X\widehat{\beta}} ; to be supplied if \code{model="logistic"}}
}
\value{
\item{proj}{The projection direction}
}
\description{
Constructs the projection direction with fixed tuning parameter in high dimensional regression
}
\examples{
n = 100
p = 400
X = matrix(sample(-2:2,n*p,replace = TRUE),nrow = n,ncol = p)
resol = 1.5
step = 3
Direction_fixedtuning(X,loading=c(1,rep(0,(p-1))),mu=sqrt(2.01*log(p)/n)*resol^{-(step-1)})
}
