% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_linear.R
\name{Direction_fixedtuning}
\alias{Direction_fixedtuning}
\title{Constructs the projection direction with fixed tuning parameter in high dimensional linear regression model}
\usage{
Direction_fixedtuning(X, loading, mu = NULL)
}
\arguments{
\item{X}{Design matrix, of dimension \eqn{n} x \eqn{p}}

\item{loading}{Loading, of length \eqn{p}}

\item{mu}{Tuning parameter in construction of projection direction}
}
\value{
\item{proj}{The projection direction}
}
\description{
Constructs the projection direction with fixed tuning parameter in high dimensional linear regression model
}
\examples{
n = 100
p = 400
X = matrix(sample(-2:2,n*p,replace = TRUE),nrow = n,ncol = p)
Direction_fixedtuning(X,loading=c(1,rep(0,(p-1))),mu=2)

}
