% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_logistic.R
\name{Direction_fixedtuning_logistic}
\alias{Direction_fixedtuning_logistic}
\title{Constructs the projection direction with fixed tuning parameter toestimate functional in high-dimensional logistic regression}
\usage{
Direction_fixedtuning_logistic(Xc, loading, mu = NULL, weight, f_prime)
}
\arguments{
\item{Xc}{Design matrix}

\item{loading}{observation vector in the linear functional}

\item{mu}{Tuning parameter in construction of projection direction}

\item{weight}{A vector of weights used in re-weighting the correction term of the estimator}

\item{f_prime}{The first derivative of the logit function at the plug-in LASSO estimate of the linear functional}
}
\value{
\item{proj}{The projection direction}
}
\description{
Constructs the projection direction with fixed tuning parameter toestimate functional in high-dimensional logistic regression
}
\examples{
X = matrix(sample(-2:2,50*400,replace = TRUE),nrow=50,ncol=400)
y = rbinom(50,1,0.5)
p <- ncol(X);
n <- nrow(X);
col.norm <- 1/sqrt((1/n)*diag(t(X)\%*\%X)+0.0001);
Xnor <- X \%*\% diag(col.norm);
fit = glmnet::cv.glmnet(Xnor, y, alpha=1,family = "binomial")
htheta <- as.vector(coef(fit, s = "lambda.min"))
support<-(abs(htheta)>0.001)
Xb <- cbind(rep(1,n),Xnor);
Xc <- cbind(rep(1,n),X);
col.norm <- c(1,col.norm);
pp <- (p+1);
xnew = c(1,rep(0,399))
loading=rep(0,pp)
loading[1]=1
loading[-1]=xnew
htheta <- htheta*col.norm;
htheta <- as.vector(htheta)
f_prime <- exp(Xc\%*\%htheta)/(1+exp(Xc\%*\%htheta))^2
Direction_fixedtuning_logistic(Xc,loading,mu=2,weight=rep(1,50),f_prime)
}
