% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Projection_Directions.R
\name{Direction_searchtuning}
\alias{Direction_searchtuning}
\title{Searches for the best step size and computes the projection direction in high dimensional regression}
\usage{
Direction_searchtuning(
  X,
  loading,
  model = "linear",
  mu = NULL,
  weight = NULL,
  deriv.vec = NULL,
  resol = 1.5,
  maxiter = 10
)
}
\arguments{
\item{X}{Design matrix, of dimension \eqn{n} x \eqn{p}}

\item{loading}{Loading, of length \eqn{p}}

\item{model}{The high dimensional regression model, either \code{linear} or \code{logistic}}

\item{mu}{The dual tuning parameter used in the construction of the projection direction (default = \code{NULL})}

\item{weight}{The weight vector of length \eqn{n}, used in correcting the plug-in estimator ; to be supplied if \code{model="logistic"} (default=NULL)}

\item{deriv.vec}{The first derivative vector of the logit function at \eqn{X\widehat{\beta}} ; to be supplied if \code{model="logistic"}}

\item{resol}{Resolution or the factor by which \code{mu} is increased/decreased to obtain the smallest \code{mu}
such that the dual optimization problem for constructing the projection direction converges (default = 1.5)}

\item{maxiter}{Maximum number of steps along which \code{mu} is increased/decreased to obtain the smallest \code{mu}
such that the dual optimization problem for constructing the projection direction converges (default = 10)}
}
\value{
\item{proj}{The projection direction}
}
\description{
Searches for the best step size and computes the projection direction in high dimensional regression
}
\examples{
n = 100
p = 400
X = matrix(sample(-2:2,n*p,replace = TRUE),nrow = n,ncol = p)
Direction_searchtuning(X,loading=c(1,rep(0,(p-1))))
}
