% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_linear.R
\name{Direction_searchtuning}
\alias{Direction_searchtuning}
\title{Constructs the projection direction with "optimal" tuning parameter for estimating functional in high-dimensional linear regression model}
\usage{
Direction_searchtuning(Xc, loading, mu = NULL, resol, maxiter)
}
\arguments{
\item{Xc}{Design matrix}

\item{loading}{observation vector in the linear functional}

\item{mu}{Tuning parameter in construction of projection direction}

\item{resol}{Resolution or the factor by which \code{mu} is increased/decreased to obtain the smallest \code{mu}
that gives convergence of the optimization problem for constructing the projection direction}

\item{maxiter}{Maximum number of steps along which \code{mu} is increased/decreased to obtain the smallest \code{mu}
that gives convergence of the optimization problem for constructing the projection direction}
}
\value{
\item{proj}{The projection direction}
}
\description{
Constructs the projection direction with "optimal" tuning parameter for estimating functional in high-dimensional linear regression model
}
\examples{
X = matrix(sample(-2:2,100*400,replace = TRUE),nrow=100,ncol=400)
Direction_searchtuning(X,loading=c(1,rep(0,399)),mu=NULL,resol=1.5,maxiter=10)
}
