% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_linear.R
\name{LF}
\alias{LF}
\title{Inference for linear functional in the high dimensional linear regression model}
\usage{
LF(
  X,
  y,
  loading,
  init.Lasso = NULL,
  lambda = NULL,
  intercept = TRUE,
  mu = NULL,
  step = NULL,
  resol = 1.5,
  maxiter = 10
)
}
\arguments{
\item{X}{Design matrix, of dimension \eqn{n} x \eqn{p}}

\item{y}{Outcome vector, of length \eqn{n}}

\item{loading}{Loading, of length \eqn{p}}

\item{init.Lasso}{Initial LASSO estimator of the regression vector (default = \code{NULL})}

\item{lambda}{The tuning parameter in construction of LASSO estimator of the regression vector (default = \code{NULL})}

\item{intercept}{Should intercept(s) be fitted (default = \code{TRUE})}

\item{mu}{The tuning parameter in construction of projection direction (default = \code{NULL})}

\item{step}{Number of steps (< \code{maxiter}) to obtain the smallest \code{mu} that gives convergence of the
optimization problem for constructing the projection direction (default = \code{NULL})}

\item{resol}{Resolution or the factor by which \code{mu} is increased/decreased to obtain the smallest \code{mu}
that gives convergence of the optimization problem for constructing the projection direction (default = 1.5)}

\item{maxiter}{Maximum number of steps along which \code{mu} is increased/decreased to obtain the smallest \code{mu}
that gives convergence of the optimization problem for constructing the projection direction (default = 10)}
}
\value{
\item{prop.est}{The bias-corrected estimator for the linear functional}
\item{sigma}{Estimate of the error variance in the linear regression model}
\item{se}{Standard error of the bias-corrected estimator}
\item{proj}{Optimal projection direction}
\item{step}{Number of steps (< \code{maxiter}) to obtain the smallest \code{mu} that gives convergence of the
optimization problem for constructing the projection direction}
\item{plug.in}{Plug-in LASSO estimator for the linear functional}
}
\description{
Computes the bias corrected estimator of linear functional \code{loading}\eqn{^{\top}\beta} for the high dimensional linear regression model \eqn{Y_i=X_i^{\top}\beta + \epsilon_i} and the corresponding standard error.
}
\examples{
n = 100
p = 400
X = matrix(sample(-2:2,n*p,replace = TRUE),nrow = n,ncol = p)
beta = (1:p)/25
y = X\%*\%beta + rnorm(n,0,1)
LF(X = X, y = y, loading = c(1,rep(0,(p-1))), intercept = TRUE)
}
\references{
\insertRef{linlin}{FIHR}
}
