% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_Inference_linear.R
\name{LF_Inference}
\alias{LF_Inference}
\title{Inference for linear functional in the high-dimensional linear regression model}
\usage{
LF_Inference(
  X,
  y,
  loading,
  init.Lasso = NULL,
  lambda = NULL,
  intercept = FALSE,
  mu = NULL,
  step = NULL,
  resol = 1.5,
  maxiter = 10
)
}
\arguments{
\item{X}{Design matrix}

\item{y}{Response variable}

\item{loading}{observation vector in the linear functional}

\item{init.Lasso}{initial LASSO estimator of the regression vector (default = \code{NULL})}

\item{lambda}{Tuning parameter in construction of LASSO estimator of the regression vector (default = \code{NULL})}

\item{intercept}{Should intercept(s) be fitted (default = \code{FALSE})}

\item{mu}{Tuning parameter in construction of projection direction (default = \code{NULL})}

\item{step}{Number of steps (< \code{maxiter}) to obtain the smallest \code{mu} that gives convergence of the
optimization problem for constructing the projection direction (default = \code{NULL})}

\item{resol}{Resolution or the factor by which \code{mu} is increased/decreased to obtain the smallest \code{mu}
that gives convergence of the optimization problem for constructing the projection direction (default = 1.5)}

\item{maxiter}{Maximum number of steps along which \code{mu} is increased/decreased to obtain the smallest \code{mu}
that gives convergence of the optimization problem for constructing the projection direction (default = 10)}
}
\value{
\item{prop.est}{The bias-corrected estimator for the linear functional}
\item{sigma}{Estimate of the error variance in the linear regression model}
\item{se}{Standard error of the bias-corrected estimator}
\item{proj}{Optimal projection direction}
\item{step}{Number of steps (< \code{maxiter}) to obtain the smallest \code{mu} that gives convergence of the
optimization problem for constructing the projection direction}
\item{plug.in}{Plug-in LASSO estimator for the linear functional}
}
\description{
Computes the bias corrected estimator of linear functional for the high-dimensional linear regression model and the corresponding standard error.
}
\examples{
X = matrix(sample(-2:2,100*400,replace = TRUE),nrow=100,ncol=400)
beta = (1:400)/25
y = X\%*\%beta + rnorm(100,0,1)
LF_Inference(X = X, y = y, loading = c(1,rep(0,399)), intercept = TRUE)
}
\references{
\insertRef{linlin}{FIHR}
}
