% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_Inference_logistic.R
\name{LF_Inference_logistic}
\alias{LF_Inference_logistic}
\title{Inference for linear functional in the high-dimensional logistic regression model}
\usage{
LF_Inference_logistic(
  X,
  y,
  loading,
  weight,
  lambda = NULL,
  intercept = FALSE,
  mu = NULL,
  step = NULL,
  resol = 1.5,
  maxiter = 6
)
}
\arguments{
\item{X}{Design matrix}

\item{y}{Response vector}

\item{loading}{observation vector in the linear functional}

\item{weight}{The vector of weights used in correcting the plug-in estimator}

\item{lambda}{Tuning parameter in construction of the LASSO estimator of the regression vector (default = \code{NULL})}

\item{intercept}{Should intercept(s) be fitted (default = \code{FALSE})}

\item{mu}{Tuning parameter in construction of the projection direction (default = \code{NULL})}

\item{step}{Number of steps (< \code{maxiter}) to obtain the smallest \code{mu} that gives convergence of the
optimization problem for constructing the projection direction (default = \code{NULL})}

\item{resol}{Resolution or the factor by which \code{mu} is increased/decreased to obtain the smallest \code{mu}
that gives convergence of the optimization problem for constructing the projection direction (default = 1.5)}

\item{maxiter}{Maximum number of steps along which \code{mu} is increased/decreased to obtain the smallest \code{mu}
that gives convergence of the optimization problem for constructing the projection direction (default = 6)}
}
\value{
\item{prop.est}{The bias corrected estimator of the linear functional}
\item{se}{Standard error of the bias-corrected estimator}
\item{proj}{Optimal projection direction}
\item{step}{Number of steps (< \code{maxiter}) to obtain the smallest \code{mu} that gives convergence of the
optimization problem for constructing the projection direction}
\item{plug.in}{Plug-in LASSO estimator of the linear functional}
\item{support}{Estimated support for the regression vector (using LASSO estimate)}
}
\description{
Computes the bias corrected estimator of the linear functional for the high-dimensional logistic regression model and the corresponding standard error.
}
\examples{
LF_Inference_logistic(X = matrix(sample(-2:2,50*300,replace = TRUE),nrow=50,ncol=300),
                      y = rbinom(50,1,0.5), loading = c(1,rep(0,299)),
                      intercept = TRUE, weight = rep(1,50))
}
\references{
\insertRef{CVXR}{FIHR}

\insertRef{glmnet1}{FIHR}

\insertRef{Matrix}{FIHR}

\insertRef{stats}{FIHR}
}
