% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_logistic.R
\name{LF_logistic}
\alias{LF_logistic}
\title{Inference for the case probability in high dimensional logistic regression}
\usage{
LF_logistic(
  X,
  y,
  xnew,
  weight = NULL,
  intercept = TRUE,
  init.Lasso = NULL,
  lambda = NULL,
  mu = NULL,
  step = NULL,
  resol = 1.5,
  maxiter = 10,
  alpha = 0.05
)
}
\arguments{
\item{X}{Design matrix, of dimension \eqn{n} x \eqn{p}}

\item{y}{Outcome vector, of length \eqn{n}}

\item{xnew}{Loading, of length \eqn{p}}

\item{weight}{The weight vector, of length \eqn{n}; if set to \code{NULL}, \eqn{w_i = h(X_i^{\top}\hat{\beta})(1-h(X_i^{\top}\hat{\beta}))} for \eqn{i=1,\ldots,n} (default = \code{NULL})}

\item{intercept}{Should intercept(s) be fitted (default = \code{TRUE})}

\item{init.Lasso}{Initial LASSO estimator of the regression vector (default = \code{NULL})}

\item{lambda}{The tuning parameter used in the construction of LASSO estimator of the regression vector (default = \code{NULL})}

\item{mu}{The dual tuning parameter used in the construction of the projection direction (default = \code{NULL})}

\item{step}{Number of steps (< \code{maxiter}) to obtain the smallest \code{mu}
such that the dual optimization problem for constructing the projection direction converges (default = \code{NULL})}

\item{resol}{The factor by which \code{mu} is increased/decreased to obtain the smallest \code{mu}
such that the dual optimization problem for constructing the projection direction converges (default = 1.5)}

\item{maxiter}{Maximum number of steps along which \code{mu} is increased/decreased to obtain the smallest \code{mu}
such that the dual optimization problem for constructing the projection direction converges (default = 10)}

\item{alpha}{Level of significance to test if the case probability is less than or qual to 0 (default = 0.05)}
}
\value{
\item{prop.est}{The bias corrected estimator of the case probability}
\item{se_linear}{The standard error of the bias-corrected estimator of the linear functional}
\item{CI}{The confidence interval for the case probability}
\item{decision}{The decision of whether the null hypothesis claiming the case probability = 0.5, is rejected (\code{decision}\eqn{=1}) or not (\code{decision}\eqn{=0})}
\item{proj}{The projection direction, of length \eqn{p}}
\item{plug.in}{The plug-in LASSO estimator of the case probability}
}
\description{
Computes the bias corrected estimator of the case probability h(\code{xnew}\eqn{^{\top}\beta}) for the high dimensional logistic regression \eqn{y|X \sim } Bernoulli(h(\eqn{X^{\top}\beta})) and the corresponding confidence interval. Here \eqn{h(z)=\frac{e^{z}}{1+e^{z}}}.
}
\examples{
A1gen <- function(rho,p){
 A1=matrix(0,p,p)
 for(i in 1:p){
   for(j in 1:p){
     A1[i,j]<-rho^(abs(i-j))
   }
 }
 A1
}
n = 100
p = 400
mu <- rep(0,p)
rho = 0.5
Cov <- (A1gen(rho,p))/2
beta <- rep(0,p)
beta[1:10] <- 0.5*c(1:10)/10
a0 = 0
set.seed(12)
loading <- MASS::mvrnorm(1,mu,Cov)
true <- FIHR:::expo(t(loading)\%*\%beta+a0)
X <- MASS::mvrnorm(n,mu,Cov)
exp_val <- X\%*\%beta+a0
prob <- exp(exp_val)/(1+exp(exp_val))
y <- rbinom(n,1,prob)
Est <- LF_logistic(X = X, y = y, xnew = loading)
}
\references{
\insertRef{linlog}{FIHR}
}
