% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_logistic.R
\name{LF_logistic}
\alias{LF_logistic}
\title{Inference for linear functional in the high dimensional logistic regression}
\usage{
LF_logistic(
  X,
  y,
  loading,
  weight,
  intercept = TRUE,
  init.Lasso = NULL,
  lambda = NULL,
  mu = NULL,
  step = NULL,
  resol = 1.5,
  maxiter = 10
)
}
\arguments{
\item{X}{Design matrix, of dimension \eqn{n} x \eqn{p}}

\item{y}{Outcome vector, of length \eqn{n}}

\item{loading}{Loading, of length \eqn{p}}

\item{weight}{The weight vector, of length \eqn{n}, used in correcting the plug-in estimator}

\item{intercept}{Should intercept(s) be fitted (default = \code{TRUE})}

\item{init.Lasso}{Initial LASSO estimator of the regression vector (default = \code{NULL})}

\item{lambda}{The tuning parameter used in the construction of LASSO estimator of the regression vector (default = \code{NULL})}

\item{mu}{The dual tuning parameter used in the construction of the projection direction (default = \code{NULL})}

\item{step}{Number of steps (< \code{maxiter}) to obtain the smallest \code{mu}
such that the dual optimization problem for constructing the projection direction converges (default = \code{NULL})}

\item{resol}{Resolution or the factor by which \code{mu} is increased/decreased to obtain the smallest \code{mu}
such that the dual optimization problem for constructing the projection direction converges (default = 1.5)}

\item{maxiter}{Maximum number of steps along which \code{mu} is increased/decreased to obtain the smallest \code{mu}
such that the dual optimization problem for constructing the projection direction converges (default = 10)}
}
\value{
\item{prop.est}{The bias corrected estimator of the linear functional}
\item{se}{The standard error of the bias-corrected estimator}
\item{proj}{The projection direction, of length \eqn{p}}
\item{plug.in}{The plug-in LASSO estimator of the linear functional}
}
\description{
Computes the bias corrected estimator of the linear functional \code{loading}\eqn{^{\top}\beta} for the high dimensional logistic regression \eqn{Y_i|X_i \sim } Bernoulli\eqn{(\frac{e^{X_i^{\top}\beta}}{1+e^{X_i^{\top}\beta}})} and the corresponding standard error.
}
\examples{
A1gen <- function(rho,p){
A1=matrix(0,p,p)
for(i in 1:p){
  for(j in 1:p){
    A1[i,j]<-rho^(abs(i-j))
  }
}
A1
}
n = 100
p = 400
mu <- rep(0,p)
rho = 0.5
Cov <- (A1gen(rho,p))/2
Cov2<-matrix(NA,nrow=p,ncol=p)
for(i in 1:p){
  for(j in 1:p){
    Cov2[i,j]<-0.5^(1+abs(i-j))
  }
}
beta <- rep(0,p)
beta[1:10] <- c(1:10)/5
X <- MASS::mvrnorm(n,mu,Cov)
exp_val <- X\%*\%beta
prob <- exp(exp_val)/(1+exp(exp_val))
y <- rbinom(n,1,prob)
loading <- MASS::mvrnorm(1,mu,Cov2)
LF_logistic(X = X, y = y, loading = loading, intercept = TRUE, weight = rep(1,n))
}
\references{
\insertRef{linlog}{FIHR}
}
