% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{print.summary.ITE}
\alias{print.summary.ITE}
\title{Printing Summarizing ITE}
\usage{
\method{print}{summary.ITE}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{An object of class `summary.ITE`, a result of a call to `summary.ITE`}

\item{digits}{The number of digits to use when printing}

\item{...}{arguments to pass down}
}
\description{
`print` method for class `summary.ITE`
}
\examples{
\dontrun{
#' ##-- Continuing the ITE(.) example:
summary(Est)
}
}
\keyword{internal}
