% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{print.summary.LF}
\alias{print.summary.LF}
\title{Printing Summarizing LF}
\usage{
\method{print}{summary.LF}(obj, digits = 4)
}
\arguments{
\item{obj}{An object of class `summary.LF`, a result of a call to `summary.LF`}

\item{digits}{The number of digits to use when printing (default=4)}
}
\description{
`print` method for class `summary.LF`
}
\examples{
\dontrun{
#' ##-- Continuing the LF(.) example:
summary(Est)
}
}
\keyword{internal}
