% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{print.summary.LF}
\alias{print.summary.LF}
\title{Printing Summarizing LF}
\usage{
\method{print}{summary.LF}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{An object of class `summary.LF`, a result of a call to `summary.LF`}

\item{digits}{The number of digits to use when printing}

\item{...}{arguments to pass down}
}
\description{
`print` method for class `summary.LF`
}
\examples{
\dontrun{
#' ##-- Continuing the LF(.) example:
summary(Est)
}
}
\keyword{internal}
