% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{print.summary.QF}
\alias{print.summary.QF}
\title{Printing ummarizing QF}
\usage{
\method{print}{summary.QF}(obj, digits = 4)
}
\arguments{
\item{obj}{An object of class `summary.QF`, a result of a call to `summary.QF`}
}
\description{
`print` method for class `summary.QF`
}
\examples{
\dontrun{
#' ##-- Continuing the QF(.) example:
summary(Est)
}
}
\keyword{internal}
