% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{print.summary.QF}
\alias{print.summary.QF}
\title{Printing ummarizing QF}
\usage{
\method{print}{summary.QF}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{An object of class `summary.QF`, a result of a call to `summary.QF`}

\item{digits}{The number of digits to use when printing}

\item{...}{arguments to pass down}
}
\description{
`print` method for class `summary.QF`
}
\examples{
\dontrun{
#' ##-- Continuing the QF(.) example:
summary(Est)
}
}
\keyword{internal}
