% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QF.R
\name{QF}
\alias{QF}
\title{Inference for quadratic functional in high-dimensional linear regression model}
\usage{
QF(
  X,
  y,
  test.set,
  A = "Sigma",
  init.Lasso = NULL,
  tau.vec = NULL,
  lambda = NULL,
  intercept = TRUE,
  mu = NULL,
  step = NULL,
  resol = 1.5,
  maxiter = 6
)
}
\arguments{
\item{X}{Design matrix, of dimension nvar(\eqn{n})xnobs(\eqn{p})}

\item{y}{Response variable, of length \eqn{n}}

\item{test.set}{set of indices, \code{G} in the quadratic functional}

\item{A}{Matrix A in the quadratic functional, of dimension \eqn{p}x\eqn{p} (either the population covariance matrix \code{Sigma} or a known matrix of suitable dimension ; default = \code{Sigma})}

\item{init.Lasso}{initial LASSO estimator for the regression vector (default = \code{NULL})}

\item{tau.vec}{Vector of enlargement factors for asymptotic variance of the bias-corrected estimator to handle super-efficiency (default = \code{NULL})}

\item{lambda}{Tuning parameter used in construction of initial LASSO estimator of the regression vector if \code{init.Lasso = NULL} (default = \code{NULL})}

\item{intercept}{Should intercept(s) be fitted (default = \code{TRUE})}

\item{mu}{Tuning parameter in construction of the projection direction (default = \code{NULL})}

\item{step}{Number of steps (< \code{maxiter}) to obtain the smallest \code{mu} that gives convergence of the
optimization problem for constructing the projection direction (default = \code{NULL})}

\item{resol}{Resolution or the factor by which \code{mu} is increased/decreased to obtain the smallest \code{mu}
that gives convergence of the optimization problem for constructing the projection direction (default = 1.5)}

\item{maxiter}{Maximum number of steps along which \code{mu} is increased/decreased to obtain the smallest \code{mu}
that gives convergence of the optimization problem for constructing the projection direction (default = 10)}
}
\value{
\item{prop.est}{The bias-corrected estimator of the quadratic functional restricted to \code{G}}
\item{sigma}{Estimate of the error variance in the linear regression model}
\item{se}{Standard error of the bias-corrected estimator}
\item{plug.in}{Plug-in LASSO estimator for the quadratic functional restricted to \code{G}}
}
\description{
Computes the bias-corrected estimator of the quadratic functional restricted to group \code{G} for the high-dimensional linear regression model and the corresponding standard error
}
\examples{
X = matrix(sample(-2:2,100*400,replace = TRUE),nrow=100,ncol=400)
beta = (1:400)/25
y = X\%*\%beta + rnorm(100,0,1)
QF(X = X, y = y, test.set=c(30:50))

}
\references{
\insertRef{grouplin}{FIHR}
}
