% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QF.R
\name{QF}
\alias{QF}
\title{Inference for quadratic functional in high dimensional linear regression}
\usage{
QF(
  X,
  y,
  G,
  Cov.weight = TRUE,
  A = NULL,
  init.Lasso = NULL,
  tau.vec = NULL,
  lambda = NULL,
  intercept = TRUE,
  mu = NULL,
  step = NULL,
  resol = 1.25,
  maxiter = 10
)
}
\arguments{
\item{X}{Design matrix, of dimension \eqn{n} x \eqn{p}}

\item{y}{Outcome vector, of length \eqn{n}}

\item{G}{The set of indices, \code{G} in \eqn{\beta_G^{\top}A\beta_G}}

\item{Cov.weight}{Logical, if set to \code{TRUE} then \code{A}\eqn{=\Sigma}, else need to provide an \code{A} (default = TRUE)}

\item{A}{The matrix A in the quadratic functional, of dimension \eqn{p}x\eqn{p}, when \code{Cov.weight = FALSE}}

\item{init.Lasso}{Initial LASSO estimator for the regression vector (default = \code{NULL})}

\item{tau.vec}{The vector of enlargement factors for asymptotic variance of the bias-corrected estimator to handle super-efficiency (default = \code{NULL})}

\item{lambda}{The tuning parameter used in the construction of initial LASSO estimator of the regression vector if \code{init.Lasso = NULL} (default = \code{NULL})}

\item{intercept}{Should intercept(s) be fitted (default = \code{TRUE})}

\item{mu}{The dual tuning parameter used in the construction of the projection direction (default = \code{NULL})}

\item{step}{Number of steps (< \code{maxiter}) to obtain the smallest \code{mu}
such that the dual optimization problem for constructing the projection direction converges (default = \code{NULL})}

\item{resol}{Resolution or the factor by which \code{mu} is increased/decreased to obtain the smallest \code{mu}
such that the dual optimization problem for constructing the projection direction converges (default = 1.5)}

\item{maxiter}{Maximum number of steps along which \code{mu} is increased/decreased to obtain the smallest \code{mu}
such that the dual optimization problem for constructing the projection direction converges (default = 10)}
}
\value{
\item{prop.est}{The bias-corrected estimator of the quadratic functional restricted to \code{G}}
\item{se}{The standard error of the bias-corrected estimator}
\item{proj}{The projection direction, of length \eqn{p}}
\item{plug.in}{The plug-in LASSO estimator for the quadratic functional restricted to \code{G}}
}
\description{
Computes the bias-corrected estimator of the quadratic functional \eqn{\beta_G^{\top}A\beta_G} for the high dimensional linear regression \eqn{Y_i = X_i^{\top}\beta + \epsilon} and the corresponding standard error.
}
\examples{
n = 100
p = 400
A1gen <- function(rho,p){
A1=matrix(0,p,p)
for(i in 1:p){
 for(j in 1:p){
   A1[i,j]<-rho^(abs(i-j))
 }
}
A1
}
mu <- rep(0,p)
mu[1:5] <- c(1:5)/5
rho = 0.5
Cov <- (A1gen(rho,p))/2
beta <- rep(0,p)
beta[1:10] <- c(1:10)/5
X <- MASS::mvrnorm(n,mu,Cov)
y = X\%*\%beta + rnorm(n)
test.set =c(30:50)
QF(X = X, y = y, G = test.set)

}
\references{
\insertRef{grouplin}{FIHR}
}
