% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary.ITE}
\alias{summary.ITE}
\title{Summarizing ITE}
\usage{
\method{summary}{ITE}(object, ...)
}
\arguments{
\item{object}{An object of class `ITE`, a result of a call to `ITE`}

\item{...}{arguments to pass down}
}
\value{
The function `summary.ITE` computes and returns a list of summary statistics.
\item{output.est}{a \eqn{ncol(loading.mat)} x 7 matrix with columns for the loading,
plugin(biased) estimators, bias-corrected estimators, its standard error, z-statistic,
corresponding (two-sided) p-value and significance stars; Each row corresponds to each loading.}
}
\description{
`summary` method for class `ITE`
}
\examples{
\dontrun{
##-- Continuing the ITE(.) example:
sEst = summary(Est)
sEst
}
}
\keyword{internal}
