% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{summary.QF}
\alias{summary.QF}
\title{Summarizing QF}
\usage{
\method{summary}{QF}(obj)
}
\arguments{
\item{obj}{An object of class `QF`, a result of a call to `QF`}
}
\value{
The function `summary.QF` computes and returns a list of summary statistics
of LF given `obj`
\item{output.est}{a \eqn{length(tau.vec)} x 7 matrix with columns for tau,
plugin(biased) estimators, bias-corrected estimators, its standard error, z-statistic,
corresponding (two-sided) p-value and significance stars; Each row corresponds to each tau.}
}
\description{
`summary` method for class `QF`
}
\examples{
\dontrun{
##-- Continuing the QF(.) example:
sEst = summary(Est)
sEst
}
}
